+++ ft_printf/definitivo/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ranavarr <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/22 17:02:29 by ranavarr          #+#    #+#             */
-/*   Updated: 2025/04/07 18:50:30 by ranavarr         ###   ########.fr       */
+/*   Updated: 2025/04/07 22:32:27 by ranavarr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,17 +25,17 @@ int				ft_printf(char const *str, ...);
 int				ft_putchar(char c);
 int				ft_print_str(char *str);
 int				ft_print_percent(void);
-int				ft_print_hex(int n, char *chars);
+int				ft_print_hex(unsigned int n, char *chars);
 int				ft_print_dec(int n);
 int				ft_print_unsigned(unsigned int n);
-int				ft_print_ptr(uintptr_t ptr);
+int				ft_print_ptr(unsigned long int ptr);
 
 //	utils
 int				ft_int_len(int i);
 char			*ft_itoa(int n, char *str);
 char			*ft_uitoa(unsigned int n, char *str);
 int				ft_unsigned_len(unsigned int s);
-int				ft_hex_len(int n);
+int				ft_hex_len(unsigned long int n);
 void			*ft_calloc(size_t nmemb, size_t size);
 void			*ft_memset(void *s, int c, size_t n);
